refactor(game): untangle lazy BASE-regen into data-access/tile-regen
`applyLazyRegen` mixed a pure regen computation with a fire-and-forget
Firestore writeback, and was the dependency that kept the read getters
(getOwnedTilesServer, getTileServer, the map queries) pinned inside data-server.

Extracted to lib/game/data-access/tile-regen.ts and split:
  - computeLazyBaseRegen(tile, player, now) — PURE, no I/O, returns the regen
    delta or null. Now unit-testable in isolation.
  - applyLazyRegen(tile, player, now, db) — pure compute + the non-awaited
    writeback (behavior unchanged: same early-returns, same update, same
    in-memory return shape).
  - applyLazyRegenBatch(tiles, player, now) — adminDbOrThrow + loop.

data-server imports all three back; `applyBaseRegen` (used only here) dropped
from its combat import. No cycle: tile-regen depends on ../combat, ../types,
@/lib/logger, data-access/{collections,db}.

Behavior-preserving. Verified: game suite 1460/1460, tsc + eslint clean.
data-server.ts 7,416 → 7,337 (7,865 → 7,337 across the Phase-4 chain).

This removes the blocker for moving the read getters into the data-access layer
(follow-up). Stacked on #1572.
